tests/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
grpcio==1.47.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-futures==1.0.0
six==1.16.0
urllib3==1.26.12
yapf==0.32.0
